backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
supabase
python-dotenv
passlib[argon2]
python-multipart
